change dnode error counter to int64_t

+++ src/inc/dnode.h
@@ -40,7 +40,7 @@ void    dnodeGetEpSetForPeer(SRpcEpSet *epSet);
 void    dnodeGetEpSetForShell(SRpcEpSet *epSet);
 int32_t dnodeGetDnodeId();
 void    dnodeGetClusterId(char *clusterId);
-int32_t dnodeGetDnodeError();
+int64_t dnodeGetDnodeError();
 void    dnodeIncDnodeError();
 
 void    dnodeUpdateEp(int32_t dnodeId, char *ep, char *fqdn, uint16_t *port);
